api/requirements.txt
fastapi
uvicorn
litellm
python-dotenv
pydantic
python-multipart
